+++ test/cliargsparser.cpp
@@ -3,7 +3,9 @@
 #include <cstring>
 
 #include "cliargsparser.h"
-#include "test-helpers.h"
+#include "test-helpers/envvar.h"
+#include "test-helpers/opts.h"
+#include "test-helpers/tempdir.h"
 
 using namespace newsboat;
