+++ test/opml.cpp
@@ -6,7 +6,8 @@
 #include "cache.h"
 #include "fileurlreader.h"
 #include "rssfeed.h"
-#include "test-helpers.h"
+#include "test-helpers/misc.h"
+#include "test-helpers/tempfile.h"
 
 using namespace newsboat;
