+++ test/opmlurlreader.cpp
@@ -1,10 +1,12 @@
 #include "opmlurlreader.h"
 
-#include <unistd.h>
+#include <fstream>
 #include <map>
+#include <unistd.h>
 
 #include "3rd-party/catch.hpp"
-#include "test-helpers.h"
+#include "test-helpers/misc.h"
+#include "test-helpers/tempfile.h"
 #include "utils.h"
 
 using namespace newsboat;
